+++ History.md
@@ -10,6 +10,9 @@
   * Fix: `INSERT ... UPDATE` with `id = LAST_INSERT_ID(id)` in MySQL
   * Fix: `Model.find({ name: { $op1, $op2 } })` object conditions with multiple operators
   * Fix: prefixing result set with qualifiers if query contains join relations and is not dispatchable
+  * Fix: arithmetic operators
+  * Fix: unary operators
+  * Fix: `Spell.$get(index)` with LIMIT
   * Docs: `Model.transaction()`
   * Docs: definition types with `index.d.ts`